backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
mongomock==4.0.0
PyJWT==2.8.0         
python-multipart==0.0.6  